+++ src/transport/encoding/raptorq.rs
@@ -8,12 +8,13 @@ use std::convert::{TryFrom, TryInto};
 use std::io::{self, ErrorKind};
 
 use blake2::{Blake2s256, Digest};
-use raptorq::ObjectTransmissionInformation;
+use safe::{SafeObjectTransmissionInformation, TransmissionInformationError};
 
 use crate::encoding::{payload::BroadcastPayload, Marshallable};
 
 mod decoder;
 mod encoder;
+mod safe;
 
 pub(crate) use decoder::RaptorQDecoder;
 pub(crate) use encoder::RaptorQEncoder;
@@ -64,25 +65,26 @@ impl<'a> ChunkedPayload<'a> {
         &self.0.gossip_frame[0..UID_SIZE]
     }
 
-    fn transmission_info(&self) -> ObjectTransmissionInformation {
+    fn transmission_info(
+        &self,
+        max_udp_len: u64,
+    ) -> Result<SafeObjectTransmissionInformation, TransmissionInformationError>
+    {
         let slice =
             &self.0.gossip_frame[UID_SIZE..(UID_SIZE + TRANSMISSION_INFO_SIZE)];
-        let transmission_info: &[u8; TRANSMISSION_INFO_SIZE] =
-            slice.try_into().expect("slice to be length 12");
-        ObjectTransmissionInformation::deserialize(transmission_info)
+        let info = SafeObjectTransmissionInformation::try_from(slice)?;
+        match info.inner.transfer_length() < max_udp_len {
+            true => Ok(info),
+            false => Err(TransmissionInformationError::TransferLengthExceeded),
+        }
     }
 
     fn encoded_chunk(&self) -> &[u8] {
         &self.0.gossip_frame[(UID_SIZE + TRANSMISSION_INFO_SIZE)..]
     }
 
-    fn safe_uid(&self) -> [u8; UID_SIZE] {
-        let mut hasher = Blake2s256::new();
-
-        let uid = &self.0.gossip_frame[0..UID_SIZE];
-        let transmission_info =
-            &self.0.gossip_frame[UID_SIZE..(UID_SIZE + TRANSMISSION_INFO_SIZE)];
-        hasher.update(uid);
+    fn uid_with_info(&self) -> [u8; UID_SIZE + TRANSMISSION_INFO_SIZE] {
+        let uid = &self.0.gossip_frame[0..UID_SIZE + TRANSMISSION_INFO_SIZE];
 
         // Why do we need transmission info?
         //
@@ -90,10 +92,14 @@ impl<'a> ChunkedPayload<'a> {
         // it is critical to decode packets.
         // Since it is sent over UDP alongside the encoded chunked bytes,
         // corrupted transmission info can be received.
-        // If the corrupted info is part of the first received chunk, no message
-        // can ever be decoded.
-        hasher.update(transmission_info);
-        hasher.finalize().into()
+        // If the corrupted info is part of the first received chunk, no
+        // message can ever be decoded.
+        //
+        // ** UPDATE:
+        // Since the correctness of an UDP packet is already guaranteed by OS
+        // checksum checks, Hashing has been removed in order to increase the
+        // decoding performance.
+        uid.try_into().expect("slice to be length 44")
     }
 }
 
@@ -101,6 +107,8 @@ impl<'a> ChunkedPayload<'a> {
 mod tests {
 
     use std::time::Instant;
+
+    use io::{BufWriter, Cursor};
 
     use super::*;
     use crate::encoding::message::Message;
@@ -110,7 +118,7 @@ mod tests {
         Configurable, Decoder, Encoder, TransportDecoder, TransportEncoder,
     };
     #[test]
-    fn test_encode() -> Result<()> {
+    fn test_encode_raptorq() -> Result<()> {
         #[cfg(not(debug_assertions))]
         let mut data = vec![0; 3_000_000];
 
@@ -162,5 +170,109 @@ mod tests {
             "Unable to decode"
         );
         Ok(())
+    }
+
+    #[test]
+    fn test_encode_raptorq_junk() -> Result<()> {
+        #[cfg(not(debug_assertions))]
+        const DATA_LEN: usize = 3_000_000;
+
+        #[cfg(debug_assertions)]
+        const DATA_LEN: usize = 100_00;
+
+        let mut data = vec![0; DATA_LEN];
+
+        for i in 0..DATA_LEN {
+            data[i] = rand::Rng::gen(&mut rand::thread_rng());
+        }
+        let peer = PeerNode::generate("192.168.0.1:666", 0)?;
+        let header = peer.to_header();
+        let payload = BroadcastPayload {
+            height: 255,
+            gossip_frame: data,
+        };
+        println!("orig payload len {}", payload.bytes()?.len());
+        let message = Message::Broadcast(header, payload);
+        let message_bytes = message.bytes()?;
+        println!("orig message len {}", message_bytes.len());
+        let start = Instant::now();
+        let encoder = TransportEncoder::configure(
+            &TransportEncoder::default_configuration(),
+        );
+        let chunks = encoder.encode(message)?;
+        println!("Encoded in: {:?}", start.elapsed());
+        println!("encoded chunks {}", chunks.len());
+        let mut decoder = TransportDecoder::configure(
+            &TransportDecoder::default_configuration(),
+        );
+        let junks_messages = 100;
+        println!("start spamming with {junks_messages} junk messages");
+        let mut decoded = None;
+        for _ in 0..junks_messages {
+            let mut gossip_frame = vec![];
+            for _ in 0..DATA_LEN {
+                gossip_frame.push(rand::Rng::gen(&mut rand::thread_rng()));
+            }
+            let msg = Message::Broadcast(
+                header,
+                BroadcastPayload {
+                    height: 255,
+                    gossip_frame,
+                },
+            );
+            if let Ok(Some(_)) = decoder.decode(msg) {
+                panic!("This should be junk data");
+            }
+        }
+        let mut i = 0;
+        let mut sizetotal = 0;
+        println!("start decoding (with additional junk messages)");
+        let start = Instant::now();
+        let mut junk = 0;
+        for chunk in chunks {
+            i = i + 1;
+            sizetotal += chunk.bytes()?.len();
+            let cloned_chunk = clone_and_corrupt_msg(&chunk)?;
+            for _ in 0..1000 {
+                let cloned_chunk = clone_and_corrupt_msg(&cloned_chunk)?;
+                if let Ok(Some(_)) = decoder.decode(cloned_chunk) {
+                    panic!("This should be junk data");
+                }
+                junk += 1;
+            }
+            if let Some(d) = decoder.decode(chunk).unwrap() {
+                decoded = Some(d);
+                println!("Decoder after {i} messages (and {junk} messages) ");
+                break;
+            }
+        }
+        println!("Decoded in: {:?}", start.elapsed());
+        println!("avg chunks size {}", sizetotal / i);
+        assert_eq!(
+            decoded.unwrap().bytes()?,
+            message_bytes,
+            "Unable to decode"
+        );
+        Ok(())
+    }
+
+    use std::io::BufReader;
+    use std::io::Read;
+    use std::io::Seek;
+    fn clone_and_corrupt_msg(message: &Message) -> Result<Message> {
+        let mut c = Cursor::new(Vec::new());
+        let mut writer = BufWriter::new(c);
+        message.marshal_binary(&mut writer)?;
+        c = writer.into_inner()?;
+        let mut bytes = vec![];
+        c.seek(std::io::SeekFrom::Start(0))?;
+        c.read_to_end(&mut bytes)?;
+        for i in 44..bytes.len() {
+            bytes[i] = rand::Rng::gen(&mut rand::thread_rng());
+        }
+        let c = Cursor::new(bytes);
+        let mut reader = BufReader::new(c);
+        let msg = Message::unmarshal_binary(&mut reader)?;
+        Ok(msg)
     }
 }