+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `max_udp_len` configuration parameter
+- Add range checks to MTU (between 1296 and 8192)
+
+### Fixed
+
+- Fix raptorQ cache default config
+- Fix ObjectTransmissionInformation deserialization
+
+### Changed
+
+- Change the EncodedChunk UUID generation
+- Change `raptorq` dependency from `1.6` to `2.0`
+
 ## [0.6.1] - 2024-04-10
 
 ### Added